refactor(utils): merge utils files
